Add an opcode for bare-this
When "$this" is used outside ObjectContext, we would create
a local to hold it. This meant an incRef and decRef for every
invocation of such a method. In addition, "return $this" would
push the local onto the stack, incrementing the refCount, and
then the RetC would decRef the ActRec's $this.

This diff adds a BareThis opcode, which is used in preference
to the local except when the local is passed by reference, or
there are dynamic variables present.

It also special-cases isset($this), is_null($this),
return $this and $this instanceof Cls, to avoid the
refcounting in those common cases.