When storing to locals, always store the type for strings
For string-typed locals, we can't skip storing the type even when the
old and new types are the same at compile time. That's because the
guards don't distinguish between static and non-static strings, so
translations should be general to handle both cases.

Also added checks for non-generic cells in SpillStack.