IR, Guard exits Jcc directly to anchor translation
Check the beginning of traces for a contiguous sequence of
Marker, DefFP, DefSP, and either LdLoc or LdStack with a label.
Record in the LdLoc and LdStack instructions that these
should branch directly to the anchor translation.
This is almost exactly what is needed to enable the
same optimization for slow exits inside the trace.