Ugly fix to the grammar to accept (string) when we want a (type)
This case can happen when a function takes a function as input:

function f1((function (string): int) $f): int {
  return $f("hello");
}

I talked to pad about "fixing" the lexer and having grammar rules for the casts. At the
end, it's going to be more work and just as ugly since we have methods called
"string" or "int".

A possible solution to these ambiguities would be to force constants to be capitalized, variable names
to start with a lower case letter, etc. I don't think such rules would play well with our goals to
be php-compatible.