Emit better mid-tracelet guards when possible
Mid-tracelet variant guards account for some nontrivial
amount of our branches to astubs. If we need to take a side exit
because of one of these and there are no dirty registers, we can emit
a very compact inline side exit. And if the current instruction
doesn't have a stack offset we can just emit a single jnz straight to
the destination.