Move a test right before a conditional jump
I surveyed all of our cmp and test instructions to make sure they come right
before the jcc or setcc instructions that they drive, because processors
apparently like it that way. This is the only one I found that has non-nops in
between. There are several examples of cmp-nop-jcc, which is due to alignment
gaps and can be fixed separately.