Add nv* virtual methods to ArrayData.
Adding implementations of the nv* virtual methods to ArrayData
lets us remove the IsHphpArray branches along the hot path in
various helpers.  They're implemented genericly in ArrayData.

Of the various ArrayData subclasses, only SharedMap and
NameValueTableWrapper are accessed frequently enough to deserve
specialized implementations (especially SharedMap).

This diff also factors out the calls to raise_notice() into
helpers in ArrayData that the various subclasses can all share.

IsHphpArray() is now only used in a few ASSERTs, so I reimplemented
it in terms of dynamic_cast.  I think we can rip it out in a future
diff, although the JIT still needs a fast is-hphp-array check to
guard inlined nvGet operations.