Get rid of alignment gaps between test and jcc
Put some more test and jcc instructions next to each other.

I discovered this oddness with emitCondJmp: it emits a jcc immediately followed
by a jmp, and it forces them to both be on the same cacheline. This could result
in more alignment gap than we need, maybe. Is this necessary for correctness, or
is it OK if the jcc and jmp have a cacheline boundary between them? We already
overwrite them separately and non-atomically anyway.