Add a "silent type"
We would like function foo(@int $x) to be equivalent to not having
any type information (i.e. equivalent to function foo($x)).

For now we are going to silently drop the type. At some point in the future,
we'll keep the type information around and log mismatches (but unlike a normal
type hint, we don't want to fatal in case of mismatch).

This kind of silent type is going to be useful to automatically migrate code
from php to hack.