Fix sp when FCallArray target is intercepted
If the target of an FCallArray was intercepted, and the
intercept handler decided to skip the original function,
we didnt update $rbx, leaving an actrec and the parameter
array on the stack.

The whole system was surprisingly robust against this. If an
exception was thrown, and the register state was dirty (the
usual case) we would recover. But if, eg, we re-entered and
threw an exception we would double free the array, and attempt
to free the ActRec as if it were Cells.