Make some JccBlock branches prediction friendly
I profiled our JccBlocks to see which ones are taken often
(see the attached task). If a JccBlock is taken the majority of the
time it's probably going to be causing a bunch of branch
mispredictions, so I've reorganized the lopsided ones to make the
forward jcc the uncommon case.