Improve function prolog guards
The prolog guards were optimized for the case of an
8 byte Func*. Now that we expect them to be 4 bytes,
optimize for that case, reducing the guard size from 23
to 14 bytes, and from 4 instructions to 2.

In addition, store the unguarded address in the Func's
prolog table; in the truly polymorphic case, when the
guard complains, we dont need to check again after
looking up the correct function in the prolog table.