Better memcpy.
Looking at the contents of icache ways at evict time, little
fragments of __memcpy_ssse3_back are all over the place. In the version
of glibc we're using, this routine is staggeringly general, and contains
carefully unrolled and tuned loops for every x86-64 civilization has ever
seen. Its final weight is around 11KB.

This is a tuned memcpy for our hardware. It assumes the presence of
movdqu instructions. Since it is statically linked into our binary, we
don't consume a TLB entry for its PLT entry.