Use a bitmask for common classes and interfaces to speed up instance checks
During warmup requests, the interpreter keeps track of the most
commonly used names for InstanceOfD and VerifyParamType. The first time one of
these instructions is translated, this information is used to assign each of
the top 127 of these names a bit in Class::m_instanceBits. This allows for
very efficient instance checking for classes and interfaces: if the
target name is in this new map, we can skip all the usual drama and just check
a bit in the object's Class struct. Since the bits are keyed on names and not
Class*s, they can be used in sandbox mode and for non-unique classes.