Fixed stack depth computation
Stack depth computation had a number of issues. This diff introduces a new
way to compute the stack depth:

- Added a new structure, StackDepth. This structure is linked to a block of
instructions (usually starting at a label), and tracks the current stack depth
in this block. This tracking can take two forms:
  * absolute depth: the depth of the stack is exactly known for this block
  * relative depth: the depth of the stack is unknown for now. We keep track
    of an offset, relative to the depth of the stack at the first instruction
    of the block

- Each Label structure contains a StackDepth structure
- An additional StackDepth structure is created at the beginning of each
function, with an absolute depth of 0.

During the parsing process, when a Jmp instruction is encountered, the
StackDepth structure for this jump becomes linked to the StackDepth structure
of the label. The absolute depth at the label can then be inferred from the
absolute depth at the jump.

Besides being able to track precisely the stack depth, these changes also
give us the following information:
- detection of the unreachable parts of the code
- detection of some bugs (for example when two Jmps with different stack
depths target the same label)