Machine code generic return helper, and minor RetC tweaks
Replaces frame_free_locals with an asm stub.  This involves a
change to the order that we destroy locals, and destroys m_this before
destroying locals.  Also changes it so we still inline returns for
functions with AttrMayUseVV, so all calls to this helper have more
than kFewLocals to free, allowing a little less branching.