Fix TestArrayForEach
I just broke this with my foreach-reset fix. This change may seem hacky but it
makes semantic sense to me: it's OK to "rewind" a continuation to put it at the
beginning (i.e. run it till the first yield if it's not been started) but not at
other times.