Don't use unary stubs in emitDecRef
Right now we jump to a register-specialized astubs block that
just jumps back to another stub that pushes a ton of regs and then
calls a destructor.  Since we're already specializing, we might as
well specialize the push code to be specific to the register allocator
state.  Still calls the unary stubs in the case that emitDecRef is
happening on astubs.