Redeclared classes should fatal
In hphpc, we just didnt bother to check; as with many
fatals, the idea was that its faster not to check, and
the code should have been tested in hhvm anyway.

In hhvm, we tried to get the rules right. But the rules
for hoisted classes are very complex with lots of edge
cases; typically you have to try to define each class twice,
ignoring fatals the first time. We were bitten by a case where
the first pass correctly didnt fatal, and the second pass
was omitted as an optimization.

To make matters worse, in the cases where hhvm didnt
fatal, it ignored later definitions of a class, while
hphpc always retained the last definition.