Generate meaningful COMPILER_ID/REPO_SCHEMA when not on git
Currently if HHVM is built outside of git, the COMPILER_ID
and REPO_SCHEMA values will effectively be constants which
means that schema changes won't be picked up properly.

This diff works around the absence of git by using the system
time (nanosecond scale) to make a "random enough" number.