Ad hoc vector optimization for arrays.
If the key is a small integer, try to avoid pulling the m_hash into cache. Guess that we're a vector-like array, and try skipping straight to m_data[key].

As part of this, I've knee-capped the (always dubious, imho) inlined array getm code. Since we're coming out ahead, I claim this worthwhile. Edwin, if this completely overlaps with your vector work I'll happily abandon.