Improve the generic freeLocalsHelper
Removes more branches from common cases in the
frameFreeLocals helper by calling to different offsets in a
generic-decref slide.  Also inlines the decref and destructor calls
instead of calling out to tvDecRefHelper.  To support this it also
always emits the varenv check (for AttrMayUseVV) and decrefs
ActRec::m_this in translateRetC translation.  For now this removes the
limit on total number of locals to do a specialized return (it leaves
the limit on number of reference counted locals).