Lock reads from the APC prime file
Thread building blocks issues -- concurrent map only uses spin locks
and calls sched_yield to pause, which just doesn't work well. If there
is any contention, it causes the system to spin like crazy.

This type of contention seems to be happening at startup when many
people are trying to use the same variables from APC. Based on profiling
I think that people are trying to read data structures that are very
array heavy, and in the handleUpdate spend a lot of time freeing it, all
while holding a writer lock.

This diff introduces a lock when you first read the piece of data from
disk. Because you hold a reader lock with the const_accessor, there's
no risk of the datastructure itself being freed. One thread is allowed
to read the object.

This also introduces a SmallLock datastructure which implements a mutex
in the space of a single int. As noted in the comments, this could be
made smaller. It could also be extended to other small locking structures
(eg, a pointer that could also serve as a lock).