Fix read-overrun in smart_realloc
If a block ended exactly at the end of a page, and then was
smart_realloc'd to a larger block, and the next page wasn't
mapped, you get a segfault.  the extra 8 bytes we copy aren't
initialized anyway, so if we don't meet these three conditions
then the overrun is harmless.

If realloc's nbytes paramter was smaller than the existing block,
the bug doesn't happen because we clamp the copied size to the
passed-in nbytes parameter.  So no write-overrun can occur.