Add more syntactic sugar to X64Assembler
Adds some new syntax for using the assembler, to try to make
it look a little more like an x64 DSEL.  Backward compatible with the
existing APIs, and stops short of any sort of massive conversions (we
can do this if it proves to be nice to use in new translations).
Converts a few small pieces of translator code to make sure it worked
and to experiment with using it.