Efficiently support switches with static string cases
This is similar to what I did for integer switches: when
every case in a switch is a literal string, the new SSWitch
instruction is used instead of a series of if/elses. At translation
time, if the input is a string and none of the cases are numeric
strings, we can turn it into a hashtable lookup. Otherwise we simulate
the code that would've been emitted by comparing the subject against
each case in the order they appeared in the source.