Add a timeout to pagelet_server_task_result + fix return code
The PHP master that calls into pagelet server currently needs to wait
indefinitely on the worker. This adds an optional timeout parameter so that the
PHP master can timeout after e.g. 10 seconds and log a descriptive message
rather than dieing with the generic message of "the request timed out after 30
seconds".

There was also a bug with the return code for partial flushed responses.
Sometimes the PHP master would receive a bogus return status like 65353. I
traced this down to the fact that the return code was never set for partial
responses, so it would be the initial value of `int rcode` on the stack.