Fix order in which classes get their traits imported
We were not enforcing that traits from base classes are imported
before importing traits in their derived classes.  This could result
in a derived class with an abstract trait method not "seeing" a
concrete definition of this method if the parent method got that
definition from a trait as well.  This would result in a "pure
virtual" fatal being thrown.