PREP_VAL needs to check its register is available
PREP_VAL tries to allocate a specific register (since it
will be used as a call argument). But there's no guarantee
the register is free. Clean/smash is not the right thing to
do, because it could be that one of the /other/ call arguments
is using that register.

So if the desired target register is not free, just use
any register, and let the usual argument shuffling take
care of getting it into the right place