Workaround a bug in HphpArray::u*sort
The array was being modified in place, while calling the
user sort functions, with the result that the user sort
function saw garbage. As a temporary fix, always duplicate
the input array, sort that, and then update the original
array reference.

A better fix would be to only do this for
the user sort cases. Even better would be to sort an array
of indices, and then update the array itself after the
user functions have been called. (task #1910931)

In addition, the new test case I wrote revealed a bug in
the translator's type inference. It failed to take account
of OpBindM's effect on its base. Also fixed here.