Kill DebuggerProxy's before its too late
The debugger keeps a map of DebuggerProxy's in Debugger::s_debugger.
When a DebuggerProxy is destroyed, it has to wait for the
corresponding thread to finish its work, and then exit. As things
stood, this didnt happen until s_debugger's destructor was called.
But that was way too late - lots of things (such as empty_array)
have been destroyed, and the thread would typically crash.

Clearing the map on Debugger::Stop shuts them down at the right
time, and things go much more smoothly.