Fix crash when 86sinit and 86pinit fail
If 86sinit or 86pinit failed (eg the reference unknown class
constants), and there's a user error handler, local
NameValueTableWrappers could be leaked into the backtrace,
resulting in use-after-free errors and crashes.

This fixes the issue by skipping these functions in the
backtrace (which is desirable anyway - they're just an
artifact of our current implementation).