Skip ActRec for builtins
A number of builtins needs to access the ActRec of the calling
function in order to get context information (this, context
class, etc). Currently they do so by explicitly skipping over
the builtin's ActRec. However, when calling builtins with
FCallBuiltin, there is no builtin ActRec, and these builtins
must be marked with the "NeedsActRec" flag.

This diff eliminates the explicit skip step, and changes it so
that the innermost ActRec is automatically skipped if it is
the builtin's ActRec. With this change, fewer builtin's need to
have the NeedsActRec flag.