Add fb_autoload_map
fb_autoload_map($map, $root) specifies a mapping
from classes, functions and constants to the files
that define them. The map has the form:

  array('class'    => array('cls' => 'cls_file.php', ...),
        'function' => array('fun' => 'fun_file.php', ...),
        'constant' => array('con' => 'con_file.php', ...),
        'failure' => callable);

If $root is non empty, it is prepended to every filename
(so will typically need to end with '/').

If the 'failure' element exists, it will be called if the
lookup in the map fails, or the file cant be included. It
takes a kind ('class', 'function' or 'constant') and the
name of the entity we're trying to autoload.

You can supply the autoload map to hphpc at compile time.
Currently, this is just used to augment hphpc's parseOnDemand
feature. It will use the map it to find more files to add to
the build based on class/function and constant names used.
This becomes necessary when all the requires are removed.