Emit more compact code for uninitting locals in prologues
This saves 2 bytes for the first local and 4 bytes for each
local after that in the inline case. Converting the loop code to the
new style saves a byte on the storel (it used to store 8 bytes to get
both _count and m_type, now it stores 4 bytes to m_type).