Add RuntimeOption::EvalJitCompareHHIR
This is meant to help compare code quality between hhir and
tx64. If hhir successfully translates a tracelet, it will be
retranslated with tx64. If the code from hhir is larger than the code
from tx64, the bytecode and disassembly for both translations will be
logged. Only the hhir tracelet is put in the SrcDB, and the cursors of
a and astubs are rewound to reuse the space from the tx64 translation.