Make separate IR instructions for guard type checks
LdLoc and LdStack, when provided a label, also do the type check, and
that's how guards were being implemented.  Although that has the nice
effect of hoisting the loads early in the trace, it's bad because the
loads effectively end up as part of the guards and they execute
redundantly when guards fail.

This diff splits the checks into their own instructions, but also
loads all the dependencies right after the guards to preserve the
hoisting effect.