Fix crash in curl_exec when exceptions are thrown
libcurl isnt exception safe. In a few places it saves a heap
pointer, replaces it with a pointer to a local, calls something
involving user callbacks, and then restores the original heap
pointer. If an exception is thrown, the original value doesnt get
restored, and when the curl handle is freed, it attempts to
free a pointer into the stack.

So we catch exceptions around the callbacks, return normally,
and then raise the exception before returning from the original
call.