Fix an issue with unused LdClsCns instructions
If you LdClsCns and don't use the output, it doesn't get
assigned a register which breaks the check for uninit.  To solve this,
split the uninit check into a separate, Essential instruction that
consumes the output of the LdClsCns.  (Presumably we could leave out
this instruction in cases where we know the class is loaded, and the
LdClsCns could rightfully be optimized out in that case.)