Fix a bug in Simplifier::simplifyCmp
When there was a non-const null and a const bool, the
simplifier would continuously flip the arguments and recurse
infinitely.  For some reason there was a check for const bools before
the check for null, although the optimizations we can do knowing the
rhs is null seem just as good as the optimizations knowing the rhs is
a const bool.  (In fact after a couple more recursive calls I think it
ends up there.)