Add SSATmp::m_type
Adds a Type field to SSATmp, so we can store the type of a variable
separate from the type hint attached to an instruction, *and* so that
we can have instructions with more than one destination, each with
a distinct type.

For the time being, this field just copies IRInstruction::m_type and
tries to keep them both up-to-date.  Future diffs need to decouple the
use cases for the two fields.