Remove TypeInstruction and some associated clean up
Gets rid of TypeInstruction, instead putting a new (possibly
Type::None) Type in the base class for the IsType instructions.  This
should go away after things are changed so SSATmps have types instead
of instructions having types (needed for multiple-dsts).  I think it'd
also be nice to remove the ExtendedInstruction vs IRInstruction
distinction (allow any instruction to have any number of srcs (and
eventually dests)); this is a small step toward that.