Fix a bug in apc serialization
ArrayData::hasInternalReference had an incorrect
short-circuit when ds was false; if it found a
Serializable object, it immediately returned false,
even though later events might have forced it to
return true.

The result is that arrays with internal references
that contain Serializable objects may not deserialize
correctly (the internal references can get split, for
example).