Reduce allocations for ImmutableMap and VectorData
ImmutableMap was doing 3 mallocs (one for the ImmutableMap
itself, one for the hash table, and one for the buckets). In
addition, it was pointlessly rounding up the number of
buckets to the next power of two. Reduce it to one malloc,
and only allocate the required number of buckets.

Similarly SharedVariant::VectorData was doing two mallocs. Reduce
to 1 by inlining the data.