Remove repo contention for TestCodeRun*.
Modify the test_code_run program to specify a distinct repo for every
test, so that no repo contention occurs.

Remove the default limit of 20 parallel tests, and for hhvm increase the
default to four times the number of CPUs to improve throughput.  hphp
uses a lot of memory when building, so leave the hphp default at the
number of CPUs.