Forget about the stack slot builtins write to.
If you line up a tracelet *just* right, you can have a live
register for the stack slot that FCallBuiltin outputs. Since FCallBuiltin
works in-memory, this register will contain a stale value. This can
cause arbitrary badness; it was first sighted by Emir as a branch that
goes the wrong way in PHP.