Remove the SpillStack vs. SpillStackAllocAR distinction
Changes guard instructions to modify the stack by returning a
new StkPtr, so we don't have to load all guarded locations with
extendStack.  This means SpillStackAllocAR doesn't ever have to do
anything differently from SpillStack (it only differs in a case where
we are spilling locations above an act rec that we loaded only as a
side-effect of guarding).  After that, move the the responsibility of
logically popping the ActRec to the Call instruction (and account for
pushing ActRecs for NewObj in getStackValue).