Remove Rematerializable from LdClsMethodCache; remove unused function
If we actually try to rematerialize a LdClsMethodCache, right
now you'll get a crash because cgLdClsMethodCache assumes the label is
non-null, but rematerialization nulls out labels.  It seems to me that
we probably don't want to be in situations where rematerializing these
happens (since they usually should just be stored to an ActRec
location), so just remove the flag rather than making it support a
null label.