Dont use a local StringData for invoke
If the StringData ever gets put into a String or Variant
it will cause the (stack) memory to be put on the GarbageList,
leading to crashes later. This was exacerbated by not incReffing
the StringData (so the String/Variant didnt even have to outlive
the StringData for this to be a problem).

We probably couldnt hit this case prior to function autoloading,
but now its easy. Making an rpc request to a non-existent function
hits it reliably.

It turns out that invoke is normally entered via the version that takes
a CStrRef, so this should actually be an improvement on the old code
(which threw away the incoming StringData and created a new one on the stack).