Fix some valgrind issues
While trying to debug a g++4.7.1 crash I came across
some valgrind warnings.

One issue was that AsyncFunc was trying to construct
an Exception, which tried to build a backtrace before
things were properly setup. Its also pretty inefficient
to generate a backtrace for every AsyncFunc which is
never used (if an exception is thrown, the backtrace
is overwritten from the thrown exception anyway).
This changes it to use an Exception* instead.