Cleaned up some load operations and added a label to LdClsCns.
Added a label to LdClsCns so that we generate better code for
it if guardStackType refines its type.

Cleaned up some of the load instructions: Removed the NR suffix from
LdMemNR, LdPropNR, and LdRefNR. Removed unused offset parameter from
LdMem. Improved docmentation of load instructions and described
rationale for having labels on some of them.

Renamed CheckUninit to CheckInit to make it consistent with CheckType
(i.e., branch to label if type is not init).