Some assertions for invalid registers
Sometimes we have PhysRegs that are InvalidReg and try to use
them anyway.  Right now, there's some cases where this ends up passing
reg::noreg to an emitC* function that treats that as a special
meaning, and you can emit machine code that didn't have the type of
memory operand you wanted.  Assert here because PhysReg stuff is
constexpr (cannot assert there).  Ideally it wouldn't implicitly
convert to Reg64 and this would all be impossible, but oh well.