Remove disabled SpillStack optimization
I think this isn't necessary now because we use GuardStk and
AssertStk to get type information into the stack.  Unless I'm missing
something, I think we shouldn't be generating any cases where we spill
something back to right where we loaded it from anymore.  (If it turns
out we are, we implement this then.)