Warn when trying to index a scalar as if it were an array
We want to produce a warning when someone tries to access a scalar as if it were an array.  So, for example:
$x = 5;
$x[0];  // Bad.
This warning should be suppressed for list assignments where the RHS is false, though, because it's common to return FALSE in place of a list when there are no more values (e.g., mysql_fetch_array).  So:
$x = FALSE;
$foo = $x[0];  // warning
list($foo) = $x;  // no warning.

To support this behavior I've introduced a new member operation to the bytecode, EQ, which is exactly like EI but is quiet about warnings.  List assignment is then translated using EQ.