Add Type::subtypeOf and SSATmp::isA()
Smith has been pointing out that most places where we do
exact type equality comparisons should probably go away.  We'd like to
be able to have types like Obj(Foo*) as a subtype of Obj (if we have
traces with known instance types), and think of ConstInt(1) as a
subtype of Int.  The intention is we'll start using isA() or
Type::subtypeOf() in most places that currently are using
Tag::operator==.  Putting this up as its own diff (pieces taken from
pending diffs by @smith and @bsimmers) since I want it in InstanceOfD.