Fix sticky eof on PlainFile
Once a PlainFile hit eof, we would never clear it - even though
we could be writing to it through a different file handle (or
another process could be writing to it).

This clears the eof flag on every read (it will immediately get
set again if the file didnt grow).