Delete atomic_cas
Relatively easy, and I didn't come across any suspicious activity while I was at
it.

There are two forms of CAS in the library: weak and strong. The difference is
that the weak version is allowed to spuriously fail; i.e. not do the CAS even
though the atomic variable has the expected value.

On x64 there is no difference in the underlying implementation, but on ARM there
is, because ARM has no atomic-CAS instruction. So the distinction now matters to
us. I've followed the guidance from the docs, which is: "When a
compare-and-exchange is in a loop, the weak version will yield better
performance on some platforms. When a weak compare-and-exchange would require a
loop and a strong one would not, the strong one is preferable." (The ARM
implementation of strong-CAS actually has a loop in it, is why.)