Fix a bug in exception handling (relating to fault regions)
When we resume after executing a fault handler via the Unwind
opcode, we were rethrowing as if the exception came from the PC
immediately following the protected region for that handler.  If that
offset was protected by a non-overlapping user catch handler, user
catch handlers (or other fault handlers) could be invoked even though
the original exception didn't come from their region.  This changes it
to track the number of nested handlers we've hit and rethrow from the
same PC that originally raised it.