change abstract function check to be a real method instead of a check at the caller
Instead of checking at every single callsite if a function is abstract, instead lets write out a method that just immediately fatals. I added an immediate to the Fatal opcode to say if the stack should be popped or not.

The hardest part of this diff was threading the `skipFrame` through everything. I ran into an ambiguous method issue and had to solve it with an Enum. I chose to create a new method instead of adding a boolean to `raise_errror` because I don't like boolean params.